requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Database
sqlalchemy>=2.0.0

# Data visualization
streamlit-aggrid>=0.3.4
altair>=5.0.0

# File handling
openpyxl>=3.1.0
xlsxwriter>=3.1.0
reportlab>=4.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Environment
python-dotenv>=1.0.0